requirements.txt
oauthlib==1.0.3
PyYAML==3.11
requests==2.9.1
requests-oauthlib==0.6.1
TwitterAPI==2.5
